+++ lib/comic_book/adapter.rb
@@ -1,6 +1,6 @@
 # NOTE: don't use ComicBook::Adapter.new directly
 #       Inherit from it when making an adapter to formats:
-#       .cb7 .cba .cbr .cbt .cbz
+#       .cb .cb7 .cba .cbr .cbt .cbz .pdf
 class ComicBook
   class Adapter
     def initialize path
@@ -11,16 +11,23 @@ class ComicBook
       raise NotImplementedError, "#{self.class} must implement #archive"
     end
 
+    def entries
+      raise NotImplementedError, "#{self.class} must implement #entries"
+    end
+
     def extract options = {}
       raise NotImplementedError, "#{self.class} must implement #extract"
     end
 
-    def pages
-      raise NotImplementedError, "#{self.class} must implement #pages"
-    end
-
     def info
       raise NotImplementedError, "#{self.class} must implement #info"
+    end
+
+    def images = entries.select(&:image?)
+    def images_and_info = entries.select { it.image? || it.info? }
+
+    def pages
+      raise NotImplementedError, "#{self.class} must implement #pages"
     end
 
     private