+++ lib/comic_book/cbr.rb
@@ -13,6 +13,10 @@ class ComicBook
       Extractor.new(path).extract options
     end
 
+    def entries
+      CLIHelpers.lsar_list(path).map { ComicBook::Entry.new it }
+    end
+
     def info
       entries = CLIHelpers.lsar_list path
       return nil unless entries.include? 'ComicInfo.xml'
@@ -24,27 +28,6 @@ class ComicBook
 
         ComicInfo.load xml_path
       end
-    end
-
-    def pages
-      CLIHelpers.lsar_list(path)
-                .select { image_file? it }
-                .map    { create_page_from_entry it }
-                .sort_by(&:name)
-    end
-
-    private
-
-    def create_page_from_entry entry
-      basename = File.basename entry
-
-      ComicBook::Page.new entry, basename
-    end
-
-    def image_file? filename
-      extension = File.extname filename.downcase
-
-      ComicBook::IMAGE_EXTENSIONS.include? extension
     end
   end
 end