+++ lib/comic_book/cbz.rb
@@ -14,6 +14,16 @@ class ComicBook
       Extractor.new(path).extract options
     end
 
+    def entries
+      names = []
+
+      Zip::File.open(path) do |zipfile|
+        zipfile.each { |entry| names << entry.name if entry.file? }
+      end
+
+      names.map { ComicBook::Entry.new it }
+    end
+
     def info
       xml = nil
 
@@ -25,32 +35,6 @@ class ComicBook
       return nil unless xml
 
       ComicInfo.load xml
-    end
-
-    def pages
-      entries = []
-
-      Zip::File.open(path) do |zipfile|
-        zipfile.each { |entry| entries << entry.name }
-      end
-
-      entries.select { image_file? it }
-             .map    { create_page_from_entry it }
-             .sort_by(&:name)
-    end
-
-    private
-
-    def create_page_from_entry entry
-      basename = File.basename entry
-
-      ComicBook::Page.new entry, basename
-    end
-
-    def image_file? filename
-      extension = File.extname filename.downcase
-
-      ComicBook::IMAGE_EXTENSIONS.include? extension
     end
   end
 end